remotememory: allow compiling on non linux systems
